refactor(distill): extract makeFakeUI and makeMockExtensionAPI factories to _test-helpers.ts
Move the fake UI factory (notifyCalls + setStatusCalls capture) and the
spy-style ExtensionAPI factory out of routing.test.ts into the shared
_test-helpers module so other consumers (notably the upcoming verify:e2e
script) can reuse them without duplicating the boilerplate.

The fake UI preserves the existing { msg, severity } notification shape,
keeping the ~10 c.msg.startsWith(...) consumer references in
routing.test.ts unchanged. routing.test.ts collapses three local
definitions (the inline UI fake, the makeUI() factory, and the
makeMockExtensionAPI() function) to a pair of imports.
